+++ test/project/uiengine/tokens/colors.js
@@ -1,8 +1,7 @@
-import path from 'path'
-import theo from 'theo'
-import { theo as UItheo } from '@uiengine/core'
+const { resolve } = require('path')
+const convert = require('@uiengine/bridge-theo')
 
-const filePath = path.resolve(__dirname, '../../lib/colors.yml')
+const filePath = resolve(__dirname, '../../lib/colors.yml')
 const titleize = string => string.replace(/([A-Z\d]+)/g, ' $1').replace(/^./, str => str.toUpperCase()).replace(/^Color /, '')
 const variablize = string => `$${string.replace(/([a-z])([A-Z\d]+)/g, '$1-$2').replace(/\s+/g, '-').toLowerCase()}`
 const modify = prop => {
@@ -13,4 +12,4 @@ const modify = prop => {
   return prop
 }
 
-module.exports = UItheo(theo).convert(filePath, modify)
+module.exports = convert(filePath, modify)